implementado products

+++ Martins.Backend.Domain/Models/Product.cs
@@ -9,10 +9,10 @@
         public decimal MaterialCost { get; set; }
         public decimal TotalCost { get; set; }
         public decimal TotalAdditionalCosts { get; set; }
-        public double StockQuantity { get; set; }
+        public decimal StockQuantity { get; set; }
         public decimal Profit { get; set; }
-        public double ProfitMarginPorcent { get; set; }
-        public double StockOnHand { get; set; }
+        public decimal ProfitMarginPorcent { get; set; }
+        public decimal StockOnHand { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
 
         public required List<ProductMaterial> BillOfMaterials { get; set; }
