+++ Martins.Backend.Domain/Models/Reports/ReportProduct.cs
@@ -11,10 +11,10 @@ namespace Martins.Backend.Domain.Models.Report
         public decimal MaterialCost { get; set; }
         public decimal TotalCost { get; set; }
         public decimal TotalAdditionalCosts { get; set; }
-        public double StockQuantity { get; set; }
+        public decimal StockQuantity { get; set; }
         public decimal Profit { get; set; }
-        public double ProfitMarginPorcent { get; set; }
-        public double StockOnHand { get; set; }
+        public decimal ProfitMarginPorcent { get; set; }
+        public decimal StockOnHand { get; set; }
         public MovementTypeProductEnum MovementType { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
